Revert "Stepper auth: add authentication capability without leaving Stepper (…"
This reverts commit e0ccb17cd5ef3d6f3a0c991d6480ad2910a8b4d0.
