--- client/landing/stepper/declarative-flow/internals/steps-repository/__user/README.md
@@ -1,3 +1,0 @@
-### User step
-
-This is the user step in Stepper. Please don't use it directly. Stepper should manage auth on your behalf. You can simply wrap the steps you want to hide behind authentication with `stepWithRequiredLogin`.
