+++ client/landing/stepper/declarative-flow/internals/steps-repository/user/style.scss
@@ -4,6 +4,7 @@ body.is-section-stepper .step-route {
 	--color-accent-60: #0e64a5;
 }
 
+
 .step-route.user {
 	display: flex;
 	flex-direction: column;
