+++ client/landing/stepper/declarative-flow/onboarding.ts
@@ -14,7 +14,6 @@ import { Flow, ProvidedDependencies } from './internals/types';
 const onboarding: Flow = {
 	name: ONBOARDING_FLOW,
 	isSignupFlow: true,
-	__experimentalUseBuiltinAuth: true,
 	useSteps() {
 		return stepsWithRequiredLogin( [
 			{
